% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_size.R
\name{grid_size}
\alias{grid_size}
\title{Grid size}
\usage{
grid_size(h, w, ss)
}
\arguments{
\item{h}{Expected height (cm) of knitting}

\item{w}{Expected width (cm) of knitting}

\item{ss}{Size square calculated from square_size()}
}
\value{
Grid size to do grid on image
}
\description{
Grid size
}
\examples{
tricot::grid_size(20, 20, square_size(31,24)) 
tricot::grid_size(5, 5, square_size(35,26))
}
