% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_size.R
\name{square_size}
\alias{square_size}
\title{Square of size}
\usage{
square_size(rg, m)
}
\arguments{
\item{rg}{Row number to have 10 cm}

\item{m}{Stitch number to have 10 cm}
}
\value{
Square size to do grid on image
}
\description{
Square of size
}
\examples{
tricot::square_size(31,24)
tricot::square_size(35,26)
}
