% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wool_weigth_needed.R
\name{wool_weigth_needed}
\alias{wool_weigth_needed}
\title{Calculation of the wool quantity needed}
\usage{
wool_weigth_needed(img_k, rg, m, p)
}
\arguments{
\item{img_k}{ggplot image from knitting_image()}

\item{rg}{Row number to have 10 cm}

\item{m}{Stitch number to have 10 cm}

\item{p}{Weight (g) to make the test square of 10*10}
}
\value{
Table with wool weight needed by color
}
\description{
Calculation of the wool quantity needed
}
\examples{

tricot::wool_weigth_needed(
  tricot::knitting_image(
  tricot::image_load(
    "img/Montagne.pdf"
    ), 
  tricot::grid_size(
    20, 
    30, 
    tricot::square_size(
      35,
      26
    )
    )
  ),
  35,
  26,
  100
)

}
