% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_load.R
\name{image_load}
\alias{image_load}
\title{Load image}
\usage{
image_load(path)
}
\arguments{
\item{path}{Path to load image}
}
\value{
Magick image in R
}
\description{
Load image
}
\examples{
tricot::image_load(here::here("img", "PeintureLaetitia.jpg"))
tricot::image_load(here::here("img", "Montagne.pdf"))
tricot::image_load(here::here("img", "Pingouin.png"))
tricot::image_load(here::here("img", "Monstre.tif"))
}
