% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting_image.R
\name{knitting_image}
\alias{knitting_image}
\title{Creation of knitting image}
\usage{
knitting_image(img, gs)
}
\arguments{
\item{img}{Magick image}

\item{gs}{Grid size calculated from grid_size()}
}
\value{
Knitting image
}
\description{
Creation of knitting image
}
\examples{

tricot::knitting_image(
  tricot::image_load(
    "img/Montagne.pdf"
    ), 
  tricot::grid_size(
    20, 
    30, 
    tricot::square_size(
      35,
      26
    )
    )
  )

}
